// src/models/informacao.ts
export interface Informacao {
   id: number;
   app: boolean;
   descricaoBoa: string;
   idUser: any;
   tipoDenuncia: string;
   dataEnvio: any;
   poluitions: string;
   status: string;
   fotoDenuncia: string;
   fotoDenunciaReferencia: string;
   nomeDenunciado: string;

    //Endereco
      latitudeDenunciado: string;
      longitudeDenunciado: string;
      ruaDenunciado: string;
      bairroDenunciado: string;
      numeroCasaDenunciado: string;
   //
   telefoneDenunciado: string;

}